pyproject.toml
[tool.poetry]
name = "drf-auth"
version = "0.1.0"
description = ""
authors = ["Sian <sianculligan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.7"
djangorestframework = "^3.12.2"
psycopg2-binary = "^2.8.6"
postgres = "^3.0.0"
djangorestframework-simplejwt = "^4.6.0"
gunicorn = "^20.0.4"
whitenoise = "^5.2.0"
httpie = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
